requirements.txt
# Core dependencies
pdfplumber==0.10.2
PyPDF2==3.0.1
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0

# NLP and text processing
nltk==3.8.1
spacy==3.8.5
langdetect==1.0.9

# Data processing
pandas==2.1.3
numpy==1.26.1

# Search functionality
whoosh==2.7.4

# Optional: OCR capabilities
pytesseract==0.3.10
Pillow==10.1.0

# PDF generation (for examples and testing)
reportlab==4.0.4
